Move Functions, Make Reporting More Consistent
Moved functions ConfigureAuth and GetEndpoint from client.go to main.go. Also changed the way several functions report errors from exiting the program to returning error values.
